UBUNTU: SAUCE: apparmor: 3.3 backport external audit data 3b3b0e4f
Instead of doing a partial revert to 3b3b0e4fc15efa507b902d90cea39e496a523c3b

Use the apparmor_audit_data.profile embedded in the lsm_audit
struct in older kernels as a pointer to the external
apparmor_audit_data.

For newer versions of apparmor we can't go back to the embedded
struct because it has a different set of fields, and changing
the embedded struct is changing the abi of the older kernel.

Signed-off-by: John Johansen <john.johansen@canonical.com>
Signed-off-by: Dongjin Kim <dongjin.kim@agreeyamobility.net>