UBUNTU: SAUCE: (no-up) apparmor: Sync to apparmor 3 - alpha 4 snapshot
This is a sync and squash to the apparmor 3 alpha 4 development snapshot
backported from 3.12. The set of patches in this squash are available
is the aa3.0-presquash branch of the dev tree.

Several of the patches in the squash have been submitted upstream.
Several more will be submitted soon, and other parts are still in
active review and development.

This squash will be updated to remove patches as they are pulled
into the upstream tree, and add new patches as they become available
to the stable apparmor dev branch.

Signed-off-by: John Johansen <john.johansen@canonical.com>
Signed-off-by: Dongjin Kim <dongjin.kim@agreeyamobility.net>